Added Several HTML Elements, Homepage Global Link and Posts Index Page Content
HTML elements datalist and embed added for the pages show and pages index pages
respectively. Datalist asks the user to rate the specific blog posts from 1 to 5
dependent upon how much they like it. Embed implemented to provide what will be a
welcoming video to the user while they choose a blog post to view, for now just dummy code
while the mp4 video itself will be added later.

Reconfigured routes file to have 'home' rather than 'root' as the application's
homepage, to then add this to the application header "Blog by Conan" so globally/
across the various pages it can be clicked so the user is always able to navigate
back to the homepage.