New Post Link Fixed, Other Revisions Made
Updated the routes to better represent the application layout.

Moved new post Simple Form code into a new form partial, and
have called it via a render in the posts/new.html.erb view page.

Posts controller updated with a new home method to allow post records
to be displayed in the home page.

Removed Tailwind code recently added to show page as it was not displaying
well in the post show page.