Routes Fix for New Post form to Display, and Index Path Adjustment
Put the resources for posts higher in the routes file,
now it is called before the individual posts routes and
therefore the form to create a new post appears when
"Create New Post" link is clicked on the homepage navbar, rather
than the link taking the user through to an already created
post page.

Blog Index link now has an index_path so it is correctly
assigned going forward.