Tailwind Cards for Blog Posts, Posts Index Reordered
Created classes for both the video (eventually will put this in), and
the individual blog posts.

Tailwind code implemented for cards within the index page. For each element
(each of the text elements for the title and body of posts) there is
individualised Tailwind styling.

This is the first step towards creating a presentable card for a blog post,
there are issues such as cards being created above the first blog post record,
and them being too wide (could be other factors such as not having long enough
blog posts however).