Bootstrap Placeholder and Select Elements Added Ahead of Future Implementation
Placeholder Animations implemented anticipating the introduction of the
About the Blog section, these are animated in that they changes colour
 and produce a loading cursor when hovered over.

Also added placeholder static elements to the index to test the difference
between the static and animated placeholder elements appearance in the about view page.

Added Select Custom CSS Bootstrap for selecting a variant from three of
 a type of blog post, the variant beeing the length. Set to disabled for
 now until it comes into use later on, without disabled code added it is
working.