Bootstrap Flash Notice Created for Post Creations
When a Blog Post is created, a flash message will appear
confirming the outcome of the create post action, originally
during the early part of this implementation it was occurring
upon the page reload, however, after inserting the flash notice
code inside the form partial and the Turbo Frame, it is now doing
the flash notice message upon the exact action of clicking "Create Post".

The notices have Bootstrap styling added to them to stand out from the
standardised Rails format of flash notices while still appearing in sync
with the style of the Bootstrap and Tailwind already existing within the
application.

Updated the application and posts controller for handling the flash notices,
the post creation can handle the info, success and error messages that are then
outputted to the user dependent upon the success or failure of the post creation.