Bootstrap Modal For Creating Blog Posts Setup, Ruby Version Tweak
Bootstrap modal has been added to the posts index page to test out how
it can be used going forward with the Blog.

Implemented Turbo in the modal and view pages to ensure it is quick to appear.

Also separately to the modal work in this branch and commit, I have
updated the Gemfile to read the Ruby version directly from the .ruby-version file.