Modal Improvement and Planning for Further Modal Work
Modal closes when the Close button is clicked but does not fully link
back to the new posts page and instead is in a static state. For now I
have added an additional button link next to it to the blog homepage so
it does not mean clicking a button leads to a dead end. Added Turbo
frame tags for this to make it seamless once working as the homepage
link is currently leading to a JavaScript error.

Modal now fades in and out on opening and closing respectively.

Elements added to the JavaScript Modal controller including for the Escape key to exit the modal.