Fix for Close and Homepage Boostrap Modal Buttons, Additional ModalT tidy up
Combined the div class for the modal and the class styling on line 2 for
the whole modal,neater in appearance to have a single div class rather
than a div class and a separate class on the same line.

Reversed placement of Close and Homepage buttons in the new post page modal.

Added data-bs-dismiss replacing data-dismiss to the Close modal button
to fix the incorrect response of only closing the modal and not enabling
interactivity with the new post page when the button is clicked.

Updated the homepage button to have attributed to it several new elements,
including role="button" aria-pressed="true" for making the button click
to work for navigating to the homepage. I am still receiving an intermittent
JavaScript content missing error so will keep this under investigation to fix going forward.