Altered Header Appearance, Navbar now Shared in Views like the Footer
Gave the Blog by Conan header a more distinctive styling, by passing in a
gray colour for the words "by Conan" while keeping "Blog" with the current
blue colour.

Navbar only previously appeared on the homepage. Moved the navbar into the
shared folder and into the previously created empty navbar partial, which
now appears throughout the various views of the application as the partial
is now also rendered in the application.html.erb layout file.