Additional Tailwind Design added to About page
Replaced About page coming soon text with a new
Tailwind About section that has a clean design with
minimal clutter. Kept the placeholder components as they
are still appearing to be useful and relevant design
components mixed in with the new Tailwind.

Text centered to emphasise the image between the two
text elements.

Will build on this new Tailwind addition further as the
blog progresses in development.

New About page image courtesy of: https://unsplash.com/photos/black-fayorit-typewriter-with-printer-paper-mk7D-4UCfmg