Ruby Implemented: Time added to Footer, Video to Posts index
Ruby time method implemented to confirm the website time as of page visit.

Implemented a Rails video helper file courtesy of 
https://gist.github.com/a-barbieri/a93281cd5c01025664e5cbd781cd917c
in order to allow video tobe embedded and playable in the posts index page.
Tested with and works with livestreams and uploaded videos.
Current video URL is for a Youtube video playlist starting with a Rails 7 Introduction Video Series by the excellent Deanin.
Added custom height and width properties for defining the video size in the webpage.